Added Dear ImGui support - tested with D3D11 and D3D12. Fixed numerous D3D12 resource-management problems as reported by d3ddebug. Moved several classes from simul::base to platform::core namespace. We should eventually remove the simul namespace entirely from Platform.

+++ DirectX12/VideoDecoder.cpp
@@ -99,7 +99,7 @@ cp::VideoDecoderResult VideoDecoder::DecodeFrame(cp::Texture* outputTexture, con
 	mGraphicsCLC.ResetCommandList();
 	mDecodeCLC.ResetCommandList();
 
-	mInputBuffer->Update(graphicsCommandList, buffer, bufferSize);
+	mInputBuffer->Update(graphicsCommandList, buffer, (uint32_t)bufferSize);
 		
 	mGraphicsCLC.ExecuteCommandList();
 	Signal(mGraphicsCLC.GetCommandQueue(), mDecodeFence);
